resume_matcher/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
sentence-transformers
spacy
keybert
scikit-learn
python-multipart
